requirements.txt
numpy
boto3
gspread
oauth2client
python-binance
